Improve AI detecting how long character has been alive for
We now use floating point precision rather than integer precision here

+++ src/characters.h
@@ -140,7 +140,7 @@ typedef struct _Character
 	/* AI */
 	int move_timer;
 	float fire_timer;
-	int time_alive;
+	float time_alive;
 } Character;
 
 extern Character gRedRover;
